[feat] Update the translation for the new webpage
