+++ website/src/locales/en.json
@@ -1,22 +1,17 @@
 {
-    "english": "English",
-    "french": "Français",
     "frenchFlag": "French flag.",
     "englishFlag": "English flag.",
     "pageNotFound": "Page not found",
     "404text": "Sorry, we couldn’t find the page you’re looking for.",
     "backHome": "Go back home",
+    "hero1": "Feeling overwhelmed by all your IT trend sources?",
+    "hero2": "Your personalized trends curator, designed to help you save time.",
+    "feature1": "An AI-powered tool to keep you perfectly informed",
+    "feature2": "Customize your email preferences to receive a newsletter tailored just for you!",
+    "feature3": "Select the news topics you want to follow",
+    "feature4": "Choose when you'd like to receive your news: daily, weekly, or on Fridays.",
+    "feature5": "Select the sources to create your personalized newsletter.",
+    "altimg": "Picture of a sample email. Each news has a title and a brief descritpion made by the AI",
     "copyright": "Copyright &copy; {{date}} Marmelab",
-    "allRightsReserved": "All rights reserved.",
-    "intro0": "An AI-powered tool to be the best informed",
-    "intro1": "Want to stay updated on your favorite topics?",
-    "intro2": "Not enough time to read all your newsletters?",
-    "intro3": "Want to save time on your research?",
-    "intro4": "CURATOR AI is made for you",
-    "person0": "Customizable",
-    "person1": "Your personalized news curator",
-    "person2": "Set your email preferences and receive a newsletter that <strong>suits you!</strong>",
-    "person3": "Choose the piece of news you want to follow",
-    "person4": "Set when you want to reveive your news : daily, weekly, on fridays...",
-    "personImg": "Picture of a sample email. Each news has a title and a brief descritpion made by the AI"
+    "allRightsReserved": "All rights reserved."
 }
