+++ website/src/locales/fr.json
@@ -1,22 +1,17 @@
 {
-    "english": "English",
-    "french": "Français",
     "frenchFlag": "Drapeau français.",
     "englishFlag": "Drapeau anglais.",
     "pageNotFound": "Page non trouvée",
-    "404text": "Désolé, nous n’avons pas pu trouver la page que vous cherchez.",
-    "backHome": "Retour à l’accueil",
+    "404text": "Désolé, nous n'avons pas pu trouver la page que vous cherchez.",
+    "backHome": "Retour à l'accueil",
+    "hero1": "Vous vous sentez submergé par toutes vos sources de tendances IT ?",
+    "hero2": "Votre curateur de tendances personnalisé, conçu pour vous faire gagner du temps.",
+    "feature1": "Un outil alimenté par l'IA pour vous tenir parfaitement informé",
+    "feature2": "Personnalisez vos préférences de messagerie pour recevoir une newsletter adaptée rien que pour vous !",
+    "feature3": "Sélectionnez les sujets d'actualité que vous souhaitez suivre",
+    "feature4": "Choisissez quand vous souhaitez recevoir vos actualités : quotidiennement, hebdomadairement ou le vendredi.",
+    "feature5": "Sélectionnez les sources pour créer votre newsletter personnalisée.",
+    "altimg": "Image d'un exemple d'email. Chaque actualité a un titre et une brève description rédigée par l'IA.",
     "copyright": "Copyright &copy; {{date}} Marmelab",
-    "allRightsReserved": "Tous droits réservés.",
-    "intro0": "Un outil alimenté par l'IA pour être le mieux informé",
-    "intro1": "Vous voulez rester informé sur vos sujets préférés ?",
-    "intro2": "Pas assez de temps pour lire toutes vos newsletters ?",
-    "intro3": "Vous voulez gagner du temps dans vos recherches ?",
-    "intro4": "CURATOR AI est fait pour vous",
-    "person0": "Personnalisable",
-    "person1": "Votre curateur de nouvelles personnalisé",
-    "person2": "Définissez vos préférences d'email et recevez une newsletter qui <strong>vous correspond !</strong>",
-    "person3": "Choisissez l'actualité que vous souhaitez suivre",
-    "person4": "Définissez quand vous souhaitez recevoir vos actualités : quotidiennement, hebdomadairement, le vendredi...",
-    "personImg": "Image d'un exemple d'email. Chaque nouvelle a un titre et une brève description rédigée par l'IA."
+    "allRightsReserved": "Tous droits réservés."
 }
